Move `SalonListViewModelAdapter` to its own file in production
