+++ NovaBlendSalon/NovaBlendSalonViewTests/SalonListViewModelAdapterTests.swift
@@ -9,25 +9,6 @@ import XCTest
 import NovaBlendSalon
 import NovaBlendSalonView
 
-class SalonListViewModelAdapter {
-    let loader: SalonLoader
-    init(loader: SalonLoader) {
-        self.loader = loader
-    }
-    
-    func load() async throws -> [SalonViewModel]  {
-        let result = try await loader.load()
-        return result.map { SalonViewModel(id: $0.id, name: $0.name, location: $0.location ?? "", phone: $0.phone ?? "", hours: prepareShopHours(openTime: $0.openTime, closeTime: $0.closeTime)) }
-    }
-    
-    private func prepareShopHours(openTime: Float, closeTime: Float) -> String{
-        guard let openTimeConverted = TimeFormatter.convertTo12HourFormat(from: openTime), let closeTimeConverted = TimeFormatter.convertTo12HourFormat(from: closeTime) else {
-            return ""
-        }
-        return "Today’s hours : \(openTimeConverted) - \(closeTimeConverted)"
-    }
-}
-
 class SpyLoader: SalonLoader {
     let salons:[NovaBlendSalon.Salon]
